+++ testapp/nested_domain/messages/events/nested_events.py
@@ -1,0 +1,8 @@
+import dataclasses
+
+from queuebie.messages import Event
+
+
+@dataclasses.dataclass(kw_only=True)
+class SomethingNestedHappened(Event):
+    other_var: int